docker_python/requirements.txt
gunicorn
falcon
psycopg2
requests
sqlparse
numpy
grpcio
grpcio-tools